+++ tests/tests_pacasam/samplers/test_targetted.py
@@ -1,15 +1,18 @@
+from math import floor
 import pytest
+from pacasam.samplers.sampler import SPLIT_COLNAME
 from pacasam.samplers.targetted import TargettedSampler
 from pacasam.utils import load_sampling_config
 
 
 def test_targetted_sampler_with_and_without_spatial_completion(synthetic_connector, session_logger):
     conf = load_sampling_config("configs/Synthetic.yml")
-    # By default: with spatial completion
+    # By default with spatial completion (Proportion of validation set is respected)
     sampler = TargettedSampler(connector=synthetic_connector, sampling_config=conf, log=session_logger)
     selection = sampler.get_patches()
     assert set(selection.sampler.unique()) == {"TargettedSampler", "SpatialSampler"}
     assert len(selection) == conf["target_total_num_patches"]
+    assert len(selection[selection[SPLIT_COLNAME] == "val"]) == floor(conf["frac_validation_set"] * conf["target_total_num_patches"])
 
     # Alternatively : without spatial completion
     sampler = TargettedSampler(connector=synthetic_connector, sampling_config=conf, log=session_logger, complete_with_spatial_sampling=False)