+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # PACASAM - CHANGELOG
 
 # 1.1.0
-- `TargettedSampler` now completes it selection to reach desired num of patches using the `SpatialSampler`.
+- `TargettedSampler` is completed by `SpatialSampler` to reach target num of patches and target validation proportion.
 
 # 1.0.0
 - `LAZExtractor` supports colorization via sources files (rgb_file + irc_file) in addition to the Géoplateforme WMS.